fix: cancel webhook on direct Stop and fix quarantine counts
Runtime.Stop now calls BeginShutdown so direct stop cancels webhook
discovery before producer waits. Stale-run quarantine increments
quarantinedExecutions instead of cleanedExecutions. Split admission
safety-floor tests under the 300-line growth cap.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
