+++ internal/cliapp/human_output.go
@@ -327,17 +327,18 @@ type runsListOutput struct {
 }
 
 type runReconcileStaleOutput struct {
-	Mode                 string   `json:"mode"`
-	CandidateRuns        int64    `json:"candidateRuns"`
-	InterruptedRuns      int64    `json:"interruptedRuns"`
-	LoopsRequeued        int64    `json:"loopsRequeued"`
-	QueueItemsRequeued   int64    `json:"queueItemsRequeued"`
-	QueueItemsCancelled  int64    `json:"queueItemsCancelled"`
-	CleanedExecutions    int64    `json:"cleanedExecutions"`
-	SkippedUncertainRuns int64    `json:"skippedUncertainRuns"`
-	RunIDs               []string `json:"runIds"`
-	LoopIDs              []string `json:"loopIds"`
-	ExecutionIDs         []string `json:"executionIds"`
+	Mode                  string   `json:"mode"`
+	CandidateRuns         int64    `json:"candidateRuns"`
+	InterruptedRuns       int64    `json:"interruptedRuns"`
+	LoopsRequeued         int64    `json:"loopsRequeued"`
+	QueueItemsRequeued    int64    `json:"queueItemsRequeued"`
+	QueueItemsCancelled   int64    `json:"queueItemsCancelled"`
+	CleanedExecutions     int64    `json:"cleanedExecutions"`
+	QuarantinedExecutions int64    `json:"quarantinedExecutions"`
+	SkippedUncertainRuns  int64    `json:"skippedUncertainRuns"`
+	RunIDs                []string `json:"runIds"`
+	LoopIDs               []string `json:"loopIds"`
+	ExecutionIDs          []string `json:"executionIds"`
 }
 
 type runOutput struct {
@@ -679,7 +680,7 @@ func writeHumanRunReconcileStale(w io.Writer, payload json.RawMessage) error {
 	if err := json.Unmarshal(payload, &data); err != nil {
 		return fmt.Errorf("decode reconcile stale response: %w", err)
 	}
-	printSection(w, "Stale runs reconciled", [][2]any{{"mode", data.Mode}, {"candidates", data.CandidateRuns}, {"interruptedRuns", data.InterruptedRuns}, {"loopsRequeued", data.LoopsRequeued}, {"queueItemsRequeued", data.QueueItemsRequeued}, {"queueItemsCancelled", data.QueueItemsCancelled}, {"cleanedExecutions", data.CleanedExecutions}, {"skippedUncertainRuns", data.SkippedUncertainRuns}, {"runIds", joinOrNone(data.RunIDs)}, {"loopIds", joinOrNone(data.LoopIDs)}, {"executionIds", joinOrNone(data.ExecutionIDs)}})
+	printSection(w, "Stale runs reconciled", [][2]any{{"mode", data.Mode}, {"candidates", data.CandidateRuns}, {"interruptedRuns", data.InterruptedRuns}, {"loopsRequeued", data.LoopsRequeued}, {"queueItemsRequeued", data.QueueItemsRequeued}, {"queueItemsCancelled", data.QueueItemsCancelled}, {"cleanedExecutions", data.CleanedExecutions}, {"quarantinedExecutions", data.QuarantinedExecutions}, {"skippedUncertainRuns", data.SkippedUncertainRuns}, {"runIds", joinOrNone(data.RunIDs)}, {"loopIds", joinOrNone(data.LoopIDs)}, {"executionIds", joinOrNone(data.ExecutionIDs)}})
 	return nil
 }
